renamed trackQuickActionTouched and moved to AnalyticsUtil + added site and post information
